[Code Health] Cleanup Sender, Receiver ownership
This patch does some significant cleanups of the Sender and
Receiver classes, in order to make embedder's lives easier (and
ideally our own) around testing and reasoning about these classes.

Sender and Receiver are now virtual abstract base classes, making
mocking easy and making it possible for tests in libcast and
tests in the embedder to not have to worry about implementation
details.

Previously, it was impossible to for embedders to appropriately
test classes they write that depend on the Receiver interface --
since you can't construct a receiver without a packet router,
and including the packet router header is banned by GN check.

Bypass-Check-License: moved some files around
Bug: 255632830
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/7594201

+++ cast/standalone_receiver/dummy_player.h
@@ -27,7 +27,7 @@ class DummyPlayer final : public Receiver::Consumer {
 
  private:
   // Receiver::Consumer implementation.
-  void OnFramesReady(int next_frame_buffer_size) final;
+  void OnFramesReady(size_t next_frame_buffer_size) final;
 
   Receiver& receiver_;
   std::vector<uint8_t> buffer_;
