+++ cast/standalone_receiver/sdl_player_base.h
@@ -94,7 +94,7 @@ class SDLPlayerBase : public Receiver::Consumer, public Decoder::Client {
   };
 
   // Receiver::Consumer implementation.
-  void OnFramesReady(int next_frame_buffer_size) final;
+  void OnFramesReady(size_t next_frame_buffer_size) final;
 
   // Determine the presentation time of the frame. Ideally, this will occur
   // based on the time progression of the media, given by the RTP timestamps.
